[P2] Absorb cli_*.py modules into phi_scan.cli package with top-level shims
