+++ phi_scan/cli_plugins.py
@@ -1,174 +1,11 @@
-"""Plugins command group — phi-scan plugins <subcommand>."""
+"""Compatibility shim — real module lives at `phi_scan.cli.plugins`.
 
-from __future__ import annotations
+This top-level re-export preserves the historical import path
+`phi_scan.cli_plugins` while the canonical home is the `phi_scan.cli`
+package. New code should import from `phi_scan.cli.plugins` directly.
+"""
 
-import json
+from phi_scan.cli import plugins as _module
+from phi_scan.cli.plugins import *  # noqa: F401,F403
 
-import typer
-from rich import box as rich_box
-from rich.markup import escape as escape_rich_markup
-from rich.table import Table
-
-from phi_scan.constants import EXIT_CODE_CLEAN
-from phi_scan.output import get_console
-from phi_scan.plugin_loader import (
-    LoadedPlugin,
-    PluginRegistry,
-    SkippedPlugin,
-    discover_plugin_registry,
-)
-
-__all__ = ["plugins_app"]
-
-plugins_app = typer.Typer(
-    name="plugins",
-    help="Discover and inspect installed recognizer plugins.",
-)
-
-# ---------------------------------------------------------------------------
-# Constants
-# ---------------------------------------------------------------------------
-
-_NO_PLUGINS_MESSAGE: str = "No recognizer plugins discovered."
-
-_STATUS_LOADED: str = "loaded"
-_STATUS_SKIPPED: str = "skipped-invalid"
-
-_TABLE_TITLE: str = "Installed Recognizer Plugins"
-_COLUMN_NAME: str = "Name"
-_COLUMN_VERSION: str = "Version"
-_COLUMN_API_VERSION: str = "API Version"
-_COLUMN_ENTITY_TYPES: str = "Entity Types"
-_COLUMN_STATUS: str = "Status"
-
-_ENTITY_TYPE_SEPARATOR: str = ", "
-_EMPTY_CELL: str = ""
-
-_JSON_KEY_PLUGINS: str = "plugins"
-_JSON_KEY_NAME: str = "name"
-_JSON_KEY_VERSION: str = "version"
-_JSON_KEY_API_VERSION: str = "api_version"
-_JSON_KEY_ENTITY_TYPES: str = "entity_types"
-_JSON_KEY_STATUS: str = "status"
-_JSON_KEY_REASON: str = "reason"
-_JSON_KEY_DISTRIBUTION: str = "distribution"
-_JSON_KEY_ENTRY_POINT: str = "entry_point"
-
-_JSON_INDENT: int = 2
-
-
-# ---------------------------------------------------------------------------
-# Public command
-# ---------------------------------------------------------------------------
-
-
-@plugins_app.command("list")
-def list_plugins(
-    is_json_output: bool = typer.Option(
-        False,
-        "--json",
-        help="Output plugin list as machine-readable JSON.",
-    ),
-) -> None:
-    """Discover installed recognizer plugins and display their metadata."""
-    registry = discover_plugin_registry()
-    if is_json_output:
-        _print_json_output(registry)
-    else:
-        _print_table_output(registry)
-    raise typer.Exit(code=EXIT_CODE_CLEAN)
-
-
-# ---------------------------------------------------------------------------
-# Table output
-# ---------------------------------------------------------------------------
-
-
-def _print_table_output(registry: PluginRegistry) -> None:
-    if not registry.loaded and not registry.skipped:
-        get_console().print(_NO_PLUGINS_MESSAGE)
-        return
-    table = _build_plugin_table(registry)
-    get_console().print(table)
-
-
-def _build_plugin_table(registry: PluginRegistry) -> Table:
-    table = Table(
-        title=_TABLE_TITLE,
-        box=rich_box.ROUNDED,
-        show_lines=True,
-    )
-    table.add_column(_COLUMN_NAME, style="bold")
-    table.add_column(_COLUMN_VERSION)
-    table.add_column(_COLUMN_API_VERSION)
-    table.add_column(_COLUMN_ENTITY_TYPES)
-    table.add_column(_COLUMN_STATUS)
-    for loaded_plugin in registry.loaded:
-        _add_loaded_row(table, loaded_plugin)
-    for skipped_plugin in registry.skipped:
-        _add_skipped_row(table, skipped_plugin)
-    return table
-
-
-def _add_loaded_row(table: Table, loaded_plugin: LoadedPlugin) -> None:
-    recognizer = loaded_plugin.recognizer
-    table.add_row(
-        recognizer.name,
-        recognizer.version,
-        recognizer.plugin_api_version,
-        _ENTITY_TYPE_SEPARATOR.join(recognizer.entity_types),
-        f"[green]{_STATUS_LOADED}[/green]",
-    )
-
-
-def _add_skipped_row(table: Table, skipped_plugin: SkippedPlugin) -> None:
-    table.add_row(
-        skipped_plugin.entry_point_name,
-        _EMPTY_CELL,
-        _EMPTY_CELL,
-        _EMPTY_CELL,
-        f"[red]{_STATUS_SKIPPED}: {escape_rich_markup(skipped_plugin.reason)}[/red]",
-    )
-
-
-# ---------------------------------------------------------------------------
-# JSON output
-# ---------------------------------------------------------------------------
-
-
-def _print_json_output(registry: PluginRegistry) -> None:
-    plugin_records = _build_json_records(registry)
-    serialized_output = json.dumps(
-        {_JSON_KEY_PLUGINS: plugin_records},
-        indent=_JSON_INDENT,
-    )
-    get_console().print(serialized_output, highlight=False)
-
-
-def _build_json_records(registry: PluginRegistry) -> list[dict[str, object]]:
-    loaded_records = [_serialize_loaded_plugin(lp) for lp in registry.loaded]
-    skipped_records = [_serialize_skipped_plugin(sp) for sp in registry.skipped]
-    return loaded_records + skipped_records
-
-
-def _serialize_loaded_plugin(loaded_plugin: LoadedPlugin) -> dict[str, object]:
-    recognizer = loaded_plugin.recognizer
-    return {
-        _JSON_KEY_NAME: recognizer.name,
-        _JSON_KEY_VERSION: recognizer.version,
-        _JSON_KEY_API_VERSION: recognizer.plugin_api_version,
-        _JSON_KEY_ENTITY_TYPES: list(recognizer.entity_types),
-        _JSON_KEY_STATUS: _STATUS_LOADED,
-        _JSON_KEY_DISTRIBUTION: loaded_plugin.distribution_name,
-        _JSON_KEY_ENTRY_POINT: loaded_plugin.entry_point_name,
-    }
-
-
-def _serialize_skipped_plugin(skipped_plugin: SkippedPlugin) -> dict[str, object]:
-    return {
-        _JSON_KEY_NAME: skipped_plugin.entry_point_name,
-        _JSON_KEY_STATUS: _STATUS_SKIPPED,
-        _JSON_KEY_REASON: skipped_plugin.reason,
-        _JSON_KEY_DISTRIBUTION: skipped_plugin.distribution_name,
-        _JSON_KEY_ENTRY_POINT: skipped_plugin.entry_point_name,
-    }
+__all__ = getattr(_module, "__all__", [name for name in dir(_module) if not name.startswith("_")])
